Lots of reworking, base for render textures
- Added low-level framebuffers and renderbuffers
- Most GL functions now take a copy of the state (State.t)
- Fixed an ID-recycling bug with all GL structures (VAO, VBO, etc...)
- Text.Fx.create temporarily takes a State.t (until we find a solution)
