+++ TODO
@@ -29,7 +29,7 @@ Interface changes:
     ☐ Alinea parameter @unsure
     ☐ Text align
     ☐ Text width
-    ☐ Factorize two modules
+    ☐ Factorize two modules - and remove State.t dependency in Text.Fx.create
     ☐ Constraint rectangle
     ☐ Transformations ? (rotation/scale/etc...)
     ☐ Set position
@@ -49,9 +49,12 @@ Interface changes:
     ☐ Check for texture max size at creation
   State:
     ☐ Rename State in GLState
-    ☐ Add state to all functions that need it, even implicitly (ie. Program.create)
+    ✔ Add state to all functions that need it, even implicitly (ie. Program.create) @done(2016-04-30 12:00) @project(Interface changes / State)
+    ✔ Fix VAO, VBO, EBO and Program ID recycling @done(2016-04-30 12:00) @project(Interface changes / State)
   Parameters:
     ☐ Add optional depth writing
+  RenderTextures:
+    ☐ Check against GL_MAX_RENDERBUFFER_SIZE when creating
     
 
 Bugs:
