+++ examples/text.ml
@@ -10,6 +10,8 @@ let settings = OgamlCore.ContextSettings.create ~msaa:8 ()
 let window =
   Window.create ~width:800 ~height:600 ~settings ~title:"Font sets tests" ()
 
+let glstate = Window.state window
+
 let font = Font.load "examples/font1.ttf"
 
 let size = 25
@@ -87,6 +89,7 @@ let random_color =
 let fxpos = Vector2f.({ x = 10. ; y = 350. })
 
 let fxtxt1 = Text.Fx.create
+  ~state:glstate
   ~text:"Awesome text!"
   ~position:fxpos
   ~font
@@ -97,6 +100,7 @@ let fxtxt1 = Text.Fx.create
 let fxpos2 = Vector2f.add (Text.Fx.advance fxtxt1) fxpos
 
 let fxtxt2 = Text.Fx.create
+  ~state:glstate
   ~text:"Success!!!"
   ~position:fxpos2
   ~font
@@ -122,6 +126,7 @@ let aa = ref false
 let draw () =
   (* Trying computing each frame *)
   let fxtxt3 = Text.Fx.create
+    ~state:glstate
     ~text:"This time we separate words a bit to check everything works."
     ~position:fxpos3
     ~font
