+++ src/graphics/stubs/program_stubs.c
@@ -15,38 +15,9 @@
 #endif
 #include <string.h>
 #include "utils.h"
+#include "types_stubs.h"
 
 #define PROGRAM(_a) (*(GLuint*) Data_custom_val(_a))
-
-
-int Val_type(GLenum type)
-{
-  switch(type)
-  {
-    case GL_INT          : return 0;
-    case GL_INT_VEC2     : return 1;
-    case GL_INT_VEC3     : return 2;
-    case GL_INT_VEC4     : return 3;
-    case GL_FLOAT        : return 4;
-    case GL_FLOAT_VEC2   : return 5;
-    case GL_FLOAT_VEC3   : return 6;
-    case GL_FLOAT_VEC4   : return 7;
-    case GL_FLOAT_MAT2   : return 8;
-    case GL_FLOAT_MAT2x3 : return 9;
-    case GL_FLOAT_MAT2x4 : return 10;
-    case GL_FLOAT_MAT3x2 : return 11;
-    case GL_FLOAT_MAT3   : return 12;
-    case GL_FLOAT_MAT3x4 : return 13;
-    case GL_FLOAT_MAT4x2 : return 14;
-    case GL_FLOAT_MAT4x3 : return 15;
-    case GL_FLOAT_MAT4   : return 16;
-    case GL_SAMPLER_1D   : return 17;
-    case GL_SAMPLER_2D   : return 18;
-    case GL_SAMPLER_3D   : return 19;
-    default: caml_failwith("Caml variant error in Val_type(1)");
-  }
-}
-
 
 void finalise_program(value v)
 {
@@ -322,7 +293,7 @@ caml_uniform_type(value id, value index)
 
   glGetActiveUniform(PROGRAM(id), Int_val(index), 0, &tmp_len, &tmp_size, &tmp_type, &tmp_name);
 
-  CAMLreturn(Val_int(Val_type(tmp_type)));
+  CAMLreturn(Val_int(Val_attrib_type(tmp_type)));
 }
 
 
@@ -341,5 +312,5 @@ caml_attribute_type(value id, value index)
 
   glGetActiveAttrib(PROGRAM(id), Int_val(index), 0, &tmp_len, &tmp_size, &tmp_type, &tmp_name);
 
-  CAMLreturn(Val_int(Val_type(tmp_type)));
+  CAMLreturn(Val_int(Val_attrib_type(tmp_type)));
 }
