+++ src/graphics/stubs/render_stubs.c
@@ -15,97 +15,7 @@
 #endif
 #include <caml/bigarray.h>
 #include "utils.h"
-
-
-GLenum Cull_val(value mode)
-{
-  switch(Int_val(mode))
-  {
-    case 0:
-      return -1;
-
-    case 1:
-      return GL_CCW;
-
-    case 2:
-      return GL_CW;
-
-    default:
-      caml_failwith("Caml variant error in Cull_val(1)");
-  }
-}
-
-
-GLenum Polygon_val(value mode)
-{
-  switch(Int_val(mode))
-  {
-    case 0:
-      return GL_POINT;
-
-    case 1:
-      return GL_LINE;
-
-    case 2:
-      return GL_FILL;
-
-    default:
-      caml_failwith("Caml variant error in Polygon_val(1)");
-  }
-}
-
-
-GLenum Depthfun_val(value fun)
-{
-  switch(Int_val(fun))
-  {
-    case 0:
-      return GL_ALWAYS;
-
-    case 1:
-      return GL_NEVER;
-
-    case 2:
-      return GL_LESS;
-
-    case 3:
-      return GL_GREATER;
-
-    case 4:
-      return GL_EQUAL;
-
-    case 5:
-      return GL_LEQUAL;
-
-    case 6:
-      return GL_GEQUAL;
-
-    case 7:
-      return GL_NOTEQUAL;
-
-    default:
-      caml_failwith("Caml variant error in Depthfun_val(1)");
-  }
-}
-
-
-value Val_error(GLenum err)
-{
-  switch(err)
-  {
-    case GL_NO_ERROR: return Val_none;
-    case GL_INVALID_ENUM      : return Val_some(Val_int(0));
-    case GL_INVALID_VALUE     : return Val_some(Val_int(1));
-    case GL_INVALID_OPERATION : return Val_some(Val_int(2));
-    case GL_INVALID_FRAMEBUFFER_OPERATION : return Val_some(Val_int(3));
-    case GL_OUT_OF_MEMORY   : return Val_some(Val_int(4));
-  #ifndef __APPLE__
-    case GL_STACK_UNDERFLOW : return Val_some(Val_int(5));
-    case GL_STACK_OVERFLOW  : return Val_some(Val_int(6));
-  #endif
-    default : return Val_none;
-  }
-}
+#include "types_stubs.h"
 
 
 // INPUT   three booleans (color, depth, stencil)
