+++ src/graphics/stubs/vao_stubs.c
@@ -16,92 +16,9 @@
 #include <caml/bigarray.h>
 #include <string.h>
 #include "utils.h"
+#include "types_stubs.h"
 
 #define VAO(_a) (*(GLuint*) Data_custom_val(_a))
-
-
-GLenum Floattype_val(value type)
-{
-  switch(Int_val(type))
-  {
-    case 0:
-      return GL_BYTE;
-
-    case 1:
-      return GL_UNSIGNED_BYTE;
-
-    case 2:
-      return GL_SHORT;
-
-    case 3:
-      return GL_UNSIGNED_SHORT;
-
-    case 4:
-      return GL_INT;
-
-    case 5:
-      return GL_UNSIGNED_INT;
-
-    case 6:
-      return GL_FLOAT;
-
-    case 7:
-      return GL_DOUBLE;
-
-    default:
-      failwith("Caml variant error in Floattype_val(1)");
-  }
-}
-
-
-GLenum Inttype_val(value type)
-{
-  switch(Int_val(type))
-  {
-    case 0:
-      return GL_BYTE;
-
-    case 1:
-      return GL_UNSIGNED_BYTE;
-
-    case 2:
-      return GL_SHORT;
-
-    case 3:
-      return GL_UNSIGNED_SHORT;
-
-    case 4:
-      return GL_INT;
-
-    case 5:
-      return GL_UNSIGNED_INT;
-
-    default:
-      failwith("Caml variant error in Inttype_val(1)");
-  }
-}
-
-
-GLenum Drawmode_val(value mode)
-{
-  switch(Int_val(mode))
-  {
-    case 0:
-      return GL_TRIANGLE_STRIP;
-
-    case 1:
-      return GL_TRIANGLE_FAN;
-
-    case 2:
-      return GL_TRIANGLES;
-
-    case 3:
-      return GL_LINES;
-
-    default:
-      failwith("Caml variant error in Drawmode_val(1)");
-  }
-}
 
 
 void finalise_vao(value v)
